refactor: 1. add rel='noreferrer noopenner' when using target='_blank' 2. remove duplicate setState in handleSendClick function 3. extract the static data from component and create a new script to store it
