+++ package.json
@@ -67,10 +67,8 @@
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.7",
     "history": "^5.3.0",
-    "i": "^0.3.7",
     "jest": "^29.6.1",
     "jest-environment-jsdom": "^29.6.1",
-    "npm": "^9.8.1",
     "postcss": "^8.4.24",
     "prettier": "^2.8.8",
     "prettier-plugin-tailwindcss": "^0.3.0",
